requirements.txt
aiogram==3.11.0
reportlab>=4.0.7
Pillow>=10.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

